log.js: Fix bytes-vs-characters issue
We can just use the size information from the server in order to find
out how many bytes we actually read, and track that separately.

There's a very minor issue remaining: it's theoretically possible that
we could upload a utf8 character split across two separate chunks, which
would cause trouble in the client.  We can fix that on the server side.

Fixes #1.