Move block logic to ChunkedUploader
This logic is a better fit here.  It also provides an opportunity to
transition the timeout management to be more explicit (and accurate).

It also means that we can upload attachments while we're waiting for
more log data (since we no longer block the mainloop until we have
"enough" data on the reader side).

Also tweak the (as yet unused) finish logic to also delete the chunks
file.

Note: we're settling into a polling-based design, and that's OK.  This
is meant to be used on a machine with tests running, so checking once
per second for a few changed files isn't going to be causing any extra
wakeups.